Cast the return value of `round()` to int
$timeRemaining is defined to be an `int`, not a `float`.

+++ src/ProgressBar.php
@@ -159,7 +159,7 @@ class ProgressBar
         if ($percent === .0 || $percent === false) {
             $timeRemaining = null;
         } else {
-            $timeRemaining = round(((1 / $percent) * $timeTaken) - $timeTaken);
+            $timeRemaining = (int) round(((1 / $percent) * $timeTaken) - $timeTaken);
         }
 
         // Poll the adapter
